# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Soccer Leg Trainer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Drill {
  id: number;
  name: string;
  /** e.g. corner, dribbling, passing, shooting, juggling */
  category: string;
  /** beginner, intermediate, advanced */
  difficulty: string;
  description: string;
  /** Coaching tips for correct form */
  keyPoints: string[];
  /** Description of the ideal leg motion pattern for AI evaluation */
  targetLegMotion: string;
  durationSeconds: number;
  /** @nullable */
  thumbnailEmoji?: string | null;
}

export interface Session {
  id: number;
  drillId: number;
  drillName: string;
  playerName: string;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  /** active, completed */
  status: string;
  /**
     * 0-100 AI score
     * @nullable
     */
  score: number | null;
  /** @nullable */
  feedbackSummary?: string | null;
  /** @nullable */
  repCount?: number | null;
}

export interface SessionInput {
  drillId: number;
  playerName: string;
}

export interface SessionUpdate {
  status?: string;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  feedbackSummary?: string | null;
  /** @nullable */
  repCount?: number | null;
}

/**
 * Keypoint data from pose detection
 */
export type FeedbackInputPoseData = {
  /** @nullable */
  kneeAngle?: number | null;
  /** @nullable */
  hipAngle?: number | null;
  /** @nullable */
  ankleFlexion?: number | null;
  /** @nullable */
  legExtension?: number | null;
  /** @nullable */
  balanceScore?: number | null;
};

export interface FeedbackInput {
  drillId: number;
  /** Keypoint data from pose detection */
  poseData: FeedbackInputPoseData;
  /** @nullable */
  repCount?: number | null;
}

export type FeedbackResultPoseQuality = {
  kneeAlignment?: string;
  hipStability?: string;
  footContact?: string;
};

export interface FeedbackResult {
  /** 0-100 performance score */
  score: number;
  /** excellent, good, needs_work, poor */
  verdict: string;
  /** Actionable coaching tips from AI */
  tips: string[];
  poseQuality: FeedbackResultPoseQuality;
}

export interface LeaderboardEntry {
  rank: number;
  playerName: string;
  totalSessions: number;
  avgScore: number;
  bestDrill: string;
}

export interface StatsSummary {
  totalSessions: number;
  avgScore: number;
  totalPlayers: number;
  topDrill: string;
  completedToday: number;
}