# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Soccer Leg Trainer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drills/{id}:
    get:
      operationId: getDrill
      tags: [drills]
      summary: Get a drill by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Drill detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drill"
        "404":
          description: Not found
...
  /sessions/{id}/feedback:
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit AI feedback for a session attempt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "201":
          description: Feedback saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResult"
...
components:
  schemas:
...
    Drill:
      type: object
      required: [id, name, category, difficulty, description, keyPoints, targetLegMotion, durationSeconds]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          description: "e.g. corner, dribbling, passing, shooting, juggling"
        difficulty:
          type: string
          description: "beginner, intermediate, advanced"
        description:
          type: string
        keyPoints:
          type: array
          items:
            type: string
          description: Coaching tips for correct form
        targetLegMotion:
          type: string
          description: Description of the ideal leg motion pattern for AI evaluation
        durationSeconds:
          type: integer
        thumbnailEmoji:
          type: ["string", "null"]
...
    FeedbackInput:
      type: object
      required: [poseData, drillId]
      properties:
        drillId:
          type: integer
        poseData:
          type: object
          description: Keypoint data from pose detection
          properties:
            kneeAngle:
              type: ["number", "null"]
            hipAngle:
              type: ["number", "null"]
            ankleFlexion:
              type: ["number", "null"]
            legExtension:
              type: ["number", "null"]
            balanceScore:
              type: ["number", "null"]
        repCount:
          type: ["integer", "null"]
    FeedbackResult:
      type: object
      required: [score, verdict, tips, poseQuality]
      properties:
        score:
          type: integer
          description: "0-100 performance score"
        verdict:
          type: string
          description: "excellent, good, needs_work, poor"
        tips:
          type: array
          items:
            type: string
          description: Actionable coaching tips from AI
        poseQuality:
          type: object
          properties:
            kneeAlignment:
              type: string
            hipStability:
              type: string
            footContact:
              type: string
...