# lib/api-zod/src/generated/types/drill.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Soccer Leg Trainer API
 * OpenAPI spec version: 0.1.0
 */

export interface Drill {
  id: number;
  name: string;
  /** e.g. corner, dribbling, passing, shooting, juggling */
  category: string;
  /** beginner, intermediate, advanced */
  difficulty: string;
  description: string;
  /** Coaching tips for correct form */
  keyPoints: string[];
  /** Description of the ideal leg motion pattern for AI evaluation */
  targetLegMotion: string;
  durationSeconds: number;
  /** @nullable */
  thumbnailEmoji?: string | null;
}