# lib/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Soccer Leg Trainer API
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackInputPoseData } from './feedbackInputPoseData';

export interface FeedbackInput {
  drillId: number;
  /** Keypoint data from pose detection */
  poseData: FeedbackInputPoseData;
  /** @nullable */
  repCount?: number | null;
}