# lib/api-zod/src/generated/types/feedbackResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Soccer Leg Trainer API
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackResultPoseQuality } from './feedbackResultPoseQuality';

export interface FeedbackResult {
  /** 0-100 performance score */
  score: number;
  /** excellent, good, needs_work, poor */
  verdict: string;
  /** Actionable coaching tips from AI */
  tips: string[];
  poseQuality: FeedbackResultPoseQuality;
}